refactor: Move mask and convert_batch into uti;s
